+++ src/TestableHttpClient/Response/JsonResponse.cs
@@ -9,7 +9,7 @@ internal class JsonResponse : ResponseBase
 
     public object? Content { get; }
 
-    protected override HttpContent? GetContent(HttpRequestMessage requestMessage)
+    protected override HttpContent? GetContent(HttpResponseContext context)
     {
         string json = JsonSerializer.Serialize(Content);
         return new StringContent(json, Encoding.UTF8, "application/json");