+++ src/TestableHttpClient/TestableHttpMessageHandler.cs
@@ -10,6 +10,7 @@ public class TestableHttpMessageHandler : HttpMessageHandler
 {
     private readonly ConcurrentQueue<HttpRequestMessage> httpRequestMessages = new ConcurrentQueue<HttpRequestMessage>();
     private IResponse response = new StatusCodeResponse(HttpStatusCode.OK);
+    private Func<HttpRequestMessage, HttpResponseMessage>? responseFactory;
 
     /// <summary>
     /// Gets the collection of captured requests made using this HttpMessageHandler.
@@ -20,11 +21,23 @@ public class TestableHttpMessageHandler : HttpMessageHandler
     {
         httpRequestMessages.Enqueue(request);
 
-        HttpResponseMessage responseMessage = await response.GetResponseAsync(request, cancellationToken).ConfigureAwait(false);
+        HttpResponseMessage responseMessage;
+        if (responseFactory is not null)
+        {
+            responseMessage = responseFactory(request);
+        }
+        else
+        {
+            responseMessage = new();
+            HttpResponseContext context = new(request, responseMessage);
+            await response.ExecuteAsync(context, cancellationToken).ConfigureAwait(false);
+        }
+
         if (responseMessage.RequestMessage is null)
         {
             responseMessage.RequestMessage = request;
         }
+
         if (responseMessage.Content is null)
         {
             responseMessage.Content = new StringContent("");
@@ -44,5 +57,25 @@ public class TestableHttpMessageHandler : HttpMessageHandler
     public void RespondWith(IResponse response)
     {
         this.response = response ?? throw new ArgumentNullException(nameof(response));
+    }
+
+    /// <summary>
+    /// Configure a factory method that creates a <see cref="HttpResponseMessage"/> that should be returned for a request.
+    /// </summary>
+    /// <param name="handler">The <see cref="TestableHttpMessageHandler"/> that should be configured.</param>
+    /// <param name="httpResponseMessageFactory">The factory method that should be called for every request. The request is passed as a parameter to the factory method and it is expected to return a HttpResponseMessage.</param>
+    /// <remarks>By default each request will receive a new response, however this is dependend on the implementation.</remarks>
+    /// <example>
+    /// testableHttpMessageHander.RespondWith(request => new ResponseMessage(HttpStatusCode.Unauthorized) { RequestMessage = request };
+    /// </example>
+    [Obsolete("Use a custom IResponse instead.")]
+    public void RespondWith(Func<HttpRequestMessage, HttpResponseMessage> httpResponseMessageFactory)
+    {
+        if (httpResponseMessageFactory is null)
+        {
+            throw new ArgumentNullException(nameof(httpResponseMessageFactory));
+        }
+
+        responseFactory = httpResponseMessageFactory;
     }
 }