+++ test/TestableHttpClient.Tests/Response/StatusCodeResponseTests.cs
@@ -10,11 +10,12 @@ public class StatusCodeResponseTests
     public async Task GetReponseAsync_WithHttpStatusCode_ReturnsCorrectStatusCode()
     {
         using HttpRequestMessage requestMessage = new();
+        using HttpResponseMessage responseMessage = new();
         StatusCodeResponse sut = new(HttpStatusCode.Continue);
 
-        var result = await sut.GetResponseAsync(requestMessage, CancellationToken.None);
+        await sut.ExecuteAsync(new HttpResponseContext(requestMessage, responseMessage), CancellationToken.None);
 
-        Assert.Equal(HttpStatusCode.Continue, result.StatusCode);
+        Assert.Equal(HttpStatusCode.Continue, responseMessage.StatusCode);
         Assert.Equal(HttpStatusCode.Continue, sut.StatusCode);
     }
 }