+++ test/TestableHttpClient.Tests/Response/TimeoutResponseTests.cs
@@ -13,7 +13,9 @@ public class TimeoutResponseTests
 
         using CancellationTokenSource cancellationTokenSource = new();
         using HttpRequestMessage requestMessage = new();
-        var task = sut.GetResponseAsync(requestMessage, cancellationTokenSource.Token);
+        using HttpResponseMessage responseMessage = new();
+        HttpResponseContext context = new(requestMessage, responseMessage);
+        var task = sut.ExecuteAsync(context, cancellationTokenSource.Token);
 
         Assert.True(cancellationTokenSource.IsCancellationRequested);
         Assert.True(task.IsCanceled);
@@ -23,7 +25,10 @@ public class TimeoutResponseTests
     public async Task GetResponseAsync_WithCancelllationTokenWithoutSource_CancelsTheTaskAsync()
     {
         TimeoutResponse sut = new();
+        using HttpRequestMessage requestMessage = new();
+        using HttpResponseMessage responseMessage = new();
+        HttpResponseContext context = new(requestMessage, responseMessage);
 
-        _ = await Assert.ThrowsAsync<TaskCanceledException>(() => sut.GetResponseAsync(new HttpRequestMessage(), CancellationToken.None));
+        _ = await Assert.ThrowsAsync<TaskCanceledException>(() => sut.ExecuteAsync(context, CancellationToken.None));
     }
 }